Add files via upload
#Chemicals-in-cosmetics

#About this dataset
Do you ever wonder what ingredients are in your favorite cosmetics? Did you know that there are over 12,000 chemicals used in cosmetics and personal care products? This dataset contains information on the chemicals used in cosmetics, including the name of the chemical, the company that manufactures it, the primary category it is used in, and the date it was first reported. So what's really in your makeup bag?

#How to use the dataset
This dataset contains information on chemicals used in cosmetics, including the name of the chemical, the company that uses it, the category of cosmetic it is used in, and the date it was first reported. The data dictionary provides more information on what each column represents. To use this dataset, you can filter by primary category, subcategory, product name, or company name to find specific information on ingredients used in cosmetics products. You can also use the 'Chemical Count' column to see how many different chemicals are present in a product.

#Project Tasks
1)Find out which chemicals were used the most in cosmetics and personal care products.

2)Find out which companies used the most reported chemicals in their cosmetics and personal care products.

3)Which brands had chemicals that were removed and discontinued? Identify the chemicals.

4)Identify the brands that had chemicals which were mostly reported in 2018.

5)Which brands had chemicals discontinued and removed?

6)Identify the period between the creation of the removed chemicals and when they were actually removed.

7)Can you tell if discontinued chemicals in bath products were removed.

8)How long were removed chemicals in baby products used? (Tip: Use creation date to tell)

9)Identify the relationship between chemicals that were mostly recently reported and discontinued. (Does most recently reported chemicals equal discontinuation of such chemicals?)

10)Identify the relationship between CSF and chemicals used in the most manufactured sub categories. (Tip: Which chemicals gave a certain type of CSF in sub categories?)

#Data Preparation and Cleaning

Explored and observed data.
Used the 'find' and 'replace' functions in Excel to remove the commas separating the values in the table. The dataset is in a CSV. file and importing it into MSSQL Server as strings was a bit of a hassle.
Inconsistences in spellings were found and corrected.
Removed duplicates.
Identify and treated missing data.